+++ gym_socks/envs/tests/test_cwh.py
@@ -39,7 +39,7 @@ class Test4DCWHSystem(unittest.TestCase):
         env = cls.env
 
         env.reset([-0.1, -0.1, 0, 0])
-        action = np.array([0, 0], dtype=np.float32)
+        action = np.array([0, 0], dtype=float)
 
         trajectory = []
         trajectory.append(env.state)
@@ -110,7 +110,7 @@ class Test6DCWHSystem(unittest.TestCase):
         env = cls.env
 
         env.reset([-0.1, -0.1, 0, 0, 0, 0])
-        action = np.array([0, 0, 0], dtype=np.float32)
+        action = np.array([0, 0, 0], dtype=float)
 
         trajectory = []
         trajectory.append(env.state)
