+++ gym_socks/envs/tests/test_nonholonomic.py
@@ -12,11 +12,11 @@ class TestNonholonomicSystem(unittest.TestCase):
     def test_corrects_angle(cls):
         system = NonholonomicVehicleEnv()
         system.disturbance_space = gym.spaces.Box(
-            low=0, high=0, shape=(3,), dtype=np.float32
+            low=0, high=0, shape=(3,), dtype=float
         )
 
         system.state = np.array([0, 0, 4 * np.pi])
-        action = np.array([0, 0], dtype=np.float32)
+        action = np.array([0, 0], dtype=float)
         obs, cost, done, _ = system.step(action)
 
         cls.assertTrue((system.state[2] <= 2 * np.pi) & (system.state[2] >= -2 * np.pi))
